스터디 작성 페이지 기능 구현
